modified controllers

+++ controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require("../models/contactModel");
 
 //@desc Get All Contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 })
 
 //@desc Get Contacts By ID
 //@route GET /api/contacts/id
-//@access public
+//@access private
 const getContactById = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
@@ -23,7 +23,7 @@ const getContactById = asyncHandler(async (req, res) => {
 
 //@desc Create Contacts
 //@route POST /api/contacts/
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
     console.log("The Request Body is", req.body);
     const { name, email, phone } = req.body;
@@ -36,18 +36,24 @@ const createContact = asyncHandler(async (req, res) => {
         name,
         email,
         phone,
+        user_id: req.user.id
     });
     res.status(201).json(contact);
 });
 
 //@desc Update Contacts
 //@route PUT /api/contacts/id
-//@access public
+//@access private
 const UpdateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact Not Found")
+    }
+
+    if (contact.user_id.toString() != req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other contacts")
     }
 
     const updateContact = await Contact.findByIdAndUpdate(req.params.id, req.body, { new: true });
@@ -56,15 +62,21 @@ const UpdateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete Contacts
 //@route DELETE /api/contacts/id
-//@access public
+//@access private
 const DeleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact Not Found");
     }
+
+    if (contact.user_id.toString() != req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other contacts")
+    }
+
     await Contact.findByIdAndDelete(req.params.id);
-    res.status(200).send("Contact Deleted Successfully");
+    res.status(200).json(contact);
 });
 
 module.exports = DeleteContact;
